+++ R/globals.R
@@ -1,4 +1,4 @@
 if(getRversion() >= "2.15.1"){
   # To remove 'no visible binding for global variable ...' in data.table commands
-  utils::globalVariables(c('t0', 'id', 'J', 'k'))
+  utils::globalVariables(c('t0', 'id', 'J', 'k', '..mycols'))
 }