+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ### Changes in Version 1.0.1 (TBD)
 
+-   Added the “cumulative percent intervened on” and “average percent
+    intervened on” to the output of the `gformula()` function
 -   Added option for users to access the bootstrap replicates of the
     parametric g-formula estimates. See the `boot_diag` argument in the
     `gformula()` function.